feat: поиск по фильмам вынесен в отдельный хук, получение фильмов в app
